Be less strict about reserved keywords
Before commit 33767856287a8fadcf3838478889909cf0c08bfb, using a thrift
reserved keyword as a fieldname threw an error, but that error was
reduced to a warning if `scroogeDisableStrict` was set. That commit
changed it to always throw an error, regardless of the value of
`scroogeDisableStrict`.

We’d like to be able to use scrooge with reserved keywords in our data,
so this commit restores the previous behaviour: using a reserved keyword
now causes an error unless `scroogeDisableStrict` is set, in which case
it causes a warning.

This reverts commit 33767856287a8fadcf3838478889909cf0c08bfb.